requirements_clean_new.txt
python-dotenv
motor
pymongo
dnspython
paramiko
matplotlib
numpy
pandas
psutil
aiohttp
aiofiles
pytz
asyncio
asyncssh
pillow
pydantic
requests
flask
py-cord
discord-py
